task(SDK-5360): Code review fixes
