+++ CleverTapSDK/CTEventBuilder.h
@@ -9,7 +9,7 @@
 
 @interface CTEventBuilder : NSObject
 
-+ (void)initializeWithValidationConfig:(CTValidationConfig*_Nullable)validationConfig;
++ (void)initializeWithValidationConfig:(CTValidationConfig* _Nonnull)validationConfig;
 
 + (void)build:(NSString * _Nonnull)eventName completionHandler:(void(^ _Nonnull )(NSDictionary* _Nullable event, NSArray<CTValidationResult*>* _Nullable errors))completion;
