+++ CleverTapSDK/Validation/CTProfileOperationType.h
@@ -1,0 +1,16 @@
+//
+//  CTProfileOperationType.h
+//  CleverTapSDK
+//
+
+typedef NS_ENUM(NSInteger, CTProfileOperation) {
+    CTProfileOperationGet = 0,
+    CTProfileOperationSet = 1,
+    CTProfileOperationRemove = 2,
+    CTProfileOperationAdd = 3,
+    CTProfileOperationIncrement = 4,
+    CTProfileOperationDecrement = 5,
+    CTProfileOperationDelete = 6,
+    CTProfileOperationArrayRemove = 7,
+    CTProfileOperationUpdate = 8
+};
